Don't hardcode archiso files anymore
This commit moves all the Anarchy files that were in archiso's airootfs
directory to their own separate directories in the root of the repo, to
prepare for removing archiso files from the repo.

A new build.sh script was created to automate copying the latest archiso
files to a temporary directory, adding anarchy's files and compiling the
new iso.