Add -k, --keep option and improve build script
When building a 32-bit ISO image on a 64-bit system, it is necessary to download some packages to the host system that will be used for the build. With the -k (or --keep) option, the build script will keep all those packages. This is useful when testing, as it avoids having to download everything again. If this option is not given, the packages will be deleted, and will have to be redownloaded when you run the script again.