Add anarchy-packages.ini and tests
Anarchy is composed of numerous interconnected scripts. A variable is in charge of storing the packages that the user wants to install. However, even with the changes added with this commit, there are package names in several scripts, which makes maintenance in case of having to verify or remove any of them extremely complex and cumbersome. With these changes a new file is added, anarchy-packages.ini, which will store variables with the package sets that are required by the installer. At the same time, a new directory is added that will store tests. In principle, check_pkgs.sh has been added, which quickly checks if any package offered by the installer has been removed from the official repos or has changed its name. In the future we intend to use this script for continuous integration or even a pre-check before the installer starts the installation.