Remove "optional software" section
Removed the last section of the installer that allowed the user to install optional software. The reason is that the system installation should be prioritized before installing extra packages. This removes a section that has given many problems to the maintainers. However, it is intended to be added in a future post-installation script.